chore: generic organization + rm some seed-enum options AB#32409
