+++ services/121-service/src/scripts/seed-program-test-one-admin.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { InterfaceScript } from '@121-service/src/scripts/scripts.module';
 import { SeedHelper } from '@121-service/src/scripts/seed-helper';
 import messageTemplateTest from '@121-service/src/seed-data/message-template/message-template-test.json';
-import organizationAdmin from '@121-service/src/seed-data/organization/organization-one-admin.json';
+import organizationGeneric from '@121-service/src/seed-data/organization/organization-generic.json';
 import programTest from '@121-service/src/seed-data/program/program-test-one-admin.json';
 
 @Injectable()
@@ -20,6 +20,6 @@ export class SeedTestOneAdmin implements InterfaceScript {
     await this.seedHelper.addOneDefaultAdminUser(program);
 
     // ***** CREATE ORGANIZATION *****
-    await this.seedHelper.addOrganization(organizationAdmin);
+    await this.seedHelper.addOrganization(organizationGeneric);
   }
 }
