--- services/121-service/src/scripts/seed-script.enum.ts
@@ -1,12 +1,0 @@
-export enum SeedScript {
-  nlrcPV = 'nlrc-pv',
-  demo = 'demo',
-  test = 'test',
-  validation = 'validation',
-  testMultiple = 'test-multiple',
-  nlrcMultiple = 'nlrc-multiple',
-  nlrcMultipleMock = 'nlrc-multiple-mock-data',
-  ethJointResponse = 'eth-joint-response',
-  krcsMultiple = 'krcs-multiple',
-  oneAdmin = 'one-admin',
-}
