--- services/121-service/src/seed-data/organization/organization-anonymous.json
@@ -1,6 +1,0 @@
-{
-  "name": "NGO A",
-  "displayName": {
-    "en": "NGO A"
-  }
-}
