--- services/121-service/src/seed-data/organization/organization-demo.json
@@ -1,10 +1,0 @@
-{
-  "name": "Demo NGO",
-  "displayName": {
-    "ar": "تجريبي NGO",
-    "en": "Demo NGO",
-    "es": "Demo NGO",
-    "fr": "NGO Démo",
-    "nl": "Demo NGO"
-  }
-}
