--- services/121-service/src/seed-data/organization/organization-eth-joint-response.json
@@ -1,6 +1,0 @@
-{
-  "name": "Ethiopia Joint Response",
-  "displayName": {
-    "en": "Ethiopia Joint Response"
-  }
-}
