+++ services/121-service/src/seed-data/organization/organization-generic.json
@@ -1,0 +1,6 @@
+{
+  "name": "NGO",
+  "displayName": {
+    "en": "NGO"
+  }
+}
