--- services/121-service/src/seed-data/organization/organization-krcs.json
@@ -1,6 +1,0 @@
-{
-  "name": "KRCS",
-  "displayName": {
-    "en": "KRCS"
-  }
-}
