--- services/121-service/src/seed-data/organization/organization-one-admin.json
@@ -1,10 +1,0 @@
-{
-  "name": "Admin NGO",
-  "displayName": {
-    "ar": "Admin NGO",
-    "en": "Admin NGO",
-    "es": "Admin NGO",
-    "fr": "NGO Admin",
-    "nl": "Admin NGO"
-  }
-}
